Fix: remove num_samples_per_class dtype

+++ train.py
@@ -108,7 +108,7 @@ def train(CONFIG):
     ######################################### Training #########################################
 
     if CONFIG.enable_open:
-        num_samples_per_class = torch.Tensor(train_dataset.sample_labels_count, dtype=torch.int).to(device)
+        num_samples_per_class = torch.Tensor(train_dataset.sample_labels_count).to(device)
         pure_noise_mean = torch.Tensor(CONFIG.pure_noise_mean).to(device)
         pure_noise_std = torch.Tensor(CONFIG.pure_noise_std).to(device)
